writing test cases for CountChar function in count.test.js file
